requirements.txt
streamlit>=1.39.0
opencv-python>=4.10.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.26.0
pillow>=10.4.0
ultralytics>=8.3.0
pydub>=0.25.1
scipy>=1.13.0
pyttsx3>=2.90
av>=12.0.0
